requirements.txt
# requirements
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
PyYAML>=6.0.0
